requirements.txt
requests
dnspython
pyfiglet
